feat: llm module
Add an llm module that allows the editor to interact with a local LLM.
We use ollama and currently there is no way to set an API key for a paid
service, but that could be added later, since the "openai" "generate"
api is already implemented.

There are two environment variables that control where Ollama is found
and which model to use:

    $ OLLAMA_HOST=localhost:11434 NOTER_OLLAMA_MODEL=gemma npm start

The OLLAMA_HOST variable is the default variable that Ollama uses
itself. It may already be set on the users computer. The
NOTER_OLLAMA_MODEL is a custom variable for our application. These
settings can move into a settings configuration file later.

Some noteable changes:

* editor.js: factored out editor "state" into: EditState, PreviewState, HelpState
  and GenerateState; the editor will scroll along, when responses are
  generated
* llm.js: module to handle LLM interaction, that is send a prompt and
  receive a response
* main.css: minor CSS tweaks
* the statusbar track total characters written by a "H"uman and
  "M"achine
* the statusbar indicates whether a model is available or not
* llm.js will fallback to a random available model, if the request model
  cannot be found

Includes basic tests using the jest test framework.