chore: try to avoid executeJavaScript
Move to ipcRenderer plus callback style messaging from main process to
renderer (e.g. setting the title of the window to the filename).

This is the beginning of a longer set of changes that will add awareness
of the current filename to the editor, so it can be better intergrated
with file open/save and navigation features.