feat: make config file location platform independent
This change allows us to use the "app" object in the
configuration/customization module and to use platform specific path as
documented here:

https://www.electronjs.org/docs/latest/api/app#appgetpathname

This required a couple of changes in main, preload as well, mostly to
accomodate the "app" object passed into the various classes.

The problem was that modules required by the initial module for electron
do not have the "app" object created properly just yet. We have to move
the initialization to a later stage.